+++ CHANGELOG.md
@@ -8,9 +8,16 @@ The following is a curated list of changes in the Enact project, newest changes 
 
 - `spotlight/SpotlightRootDecorator` to use a functional component
 
+### Changed
+
+- `ui/VirtualList.VirtualListBasic` scroll animation method for `scrollMode: native`
+
 ### Fixed
 
 - `ui/Scroller` jumping back when scrolled by long press
+- `ui/VirtualList` overscroll effect at the bottom of the list
+- `ui/VirtualList.VirtualList` overscroll effect at the bottom of the list
+- `ui/VirtualList.VirtualGridList` overscroll effect at the bottom of the list
 
 ## [5.5.0] - 2026-05-08